project11 - CompilationEngine 구현 파일 분할
